Refactor dialog components and update imports
