+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/screens/ZoneScreen.kt
@@ -56,7 +56,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.skylake.skytv.jgorunner.R
 import com.skylake.skytv.jgorunner.data.SkySharedPref
-import com.skylake.skytv.jgorunner.ui.components.ModeSelectionDialog2
+import com.skylake.skytv.jgorunner.ui.tvhome.components.TvScreenMenu
 import com.skylake.skytv.jgorunner.ui.tvhome.Main_Layout
 import com.skylake.skytv.jgorunner.ui.tvhome.Recent_Layout
 import com.skylake.skytv.jgorunner.ui.tvhome.Main_Layout_3rd
@@ -222,7 +222,7 @@ fun ZoneScreen(context: Context, onNavigate: (String) -> Unit) {
     }
 
     // Mode Dialog
-    ModeSelectionDialog2(
+    TvScreenMenu(
         showDialog = showModeDialog,
         onDismiss = { showModeDialog = false },
         onReset = {
