+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/tvhome/components/ProcessingDialog.kt
@@ -1,4 +1,4 @@
-package com.skylake.skytv.jgorunner.ui.components
+package com.skylake.skytv.jgorunner.ui.tvhome.components
 
 import android.content.Context
 import android.util.Log
@@ -12,19 +12,16 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import com.google.gson.Gson
 import com.skylake.skytv.jgorunner.data.SkySharedPref
-import com.skylake.skytv.jgorunner.ui.tvhome.M3UChannelExp
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import okhttp3.OkHttpClient
 import okhttp3.Request
 
-
-
 @Composable
-fun ProcessingDialogExp(
+fun ProcessingDialog(
     context: Context,
-    onComplete: (List<M3UChannelExp>) -> Unit,
+    onComplete: (List<M3UChannel>) -> Unit,
     onError: (String) -> Unit
 ) {
     val preferenceManager = SkySharedPref.getInstance(context)
@@ -45,14 +42,14 @@ fun ProcessingDialogExp(
             }
             try {
                 // Fetch the M3U file
-                val response = withContext(Dispatchers.IO) {
+                    val (body, response) = withContext(Dispatchers.IO) {
                     val client = OkHttpClient()
                     val request = Request.Builder().url(m3uUrl).build()
                     Log.d("ProcessingDialog", "Fetching M3U file from URL...")
-                    client.newCall(request).execute()
+                    val response = client.newCall(request).execute()
+                    val body = response.body?.string()
+                    Pair(body, response)
                 }
-
-                val body = response.body?.string()
 
                 if (!response.isSuccessful || body == null) {
                     message = "Failed to fetch M3U file. HTTP error: ${response.code}"
@@ -69,8 +66,10 @@ fun ProcessingDialogExp(
                     return@launch
                 }
                 Log.d("ProcessingDialog", "M3U file fetched successfully. Parsing...")
-                val channels = parseM3Uexp(body)
+                // Parse the M3U data
+                val channels = parseM3U(body)
                 Log.d("ProcessingDialog", "Parsed ${channels.size} channels from M3U.")
+                // Save as JSON for later use
                 val gson = Gson()
                 val json = gson.toJson(channels)
                 preferenceManager.myPrefs.channelListJson = json
@@ -108,30 +107,31 @@ fun ProcessingDialogExp(
     }
 }
 
-fun parseM3Uexp(m3uContent: String): List<M3UChannelExp> {
+// Data class for parsed channels
+data class M3UChannel(
+    val name: String,
+    val url: String,
+    val logo: String?
+)
+
+// M3U parsing function with logs
+fun parseM3U(m3uContent: String): List<M3UChannel> {
     val lines = m3uContent.lines()
-    val channels = mutableListOf<M3UChannelExp>()
+    val channels = mutableListOf<M3UChannel>()
     var i = 0
     while (i < lines.size) {
         val line = lines[i]
         if (line.startsWith("#EXTINF")) {
-            // Regex to extract key-value attributes
-            val name = Regex(",\\s*(.+)$").find(line)?.groupValues?.get(1)?.trim() ?: ""
+            val name = Regex(",\\s*(.+)$").find(line)?.groupValues?.get(1) ?: ""
             val logo = Regex("""tvg-logo="([^"]*)"""").find(line)?.groupValues?.get(1)
-            // New regex to extract group-title
-            val category = Regex("""group-title="([^"]*)"""").find(line)?.groupValues?.get(1)
-
             val url = lines.getOrNull(i + 1)?.trim() ?: ""
-
-            if (name.isNotEmpty() && url.isNotEmpty()) {
-                Log.d("parseM3Uexp", "Parsed channel: name=$name, url=$url, logo=$logo, category=$category")
-                channels.add(M3UChannelExp(name, url, logo, category))
-            }
+            Log.d("parseM3U", "Parsed channel: name=$name, url=$url, logo=$logo")
+            channels.add(M3UChannel(name, url, logo))
             i += 2
         } else {
             i++
         }
     }
-    Log.d("parseM3Uexp", "Total channels parsed: ${channels.size}")
+    Log.d("parseM3U", "Total channels parsed: ${channels.size}")
     return channels
 }
